Revert "Support bool/int8/int16/int32/int64 castings for ORC reading. (#6273)"
This reverts commit 1e2a074fa681b9f80166f87d209b45b8d138e49e.
