+++ sql-plugin/src/main/scala/com/nvidia/spark/rapids/GpuOrcScan.scala
@@ -35,7 +35,7 @@ import com.google.protobuf.CodedOutputStream
 import com.nvidia.spark.rapids.GpuMetric._
 import com.nvidia.spark.rapids.RapidsPluginImplicits._
 import com.nvidia.spark.rapids.SchemaUtils._
-import com.nvidia.spark.rapids.shims.{OrcCastingShims, OrcReadingShims, OrcShims, ShimFilePartitionReaderFactory}
+import com.nvidia.spark.rapids.shims.{OrcReadingShims, OrcShims, ShimFilePartitionReaderFactory}
 import org.apache.hadoop.conf.Configuration
 import org.apache.hadoop.fs.{FileSystem, Path}
 import org.apache.hadoop.hive.common.io.DiskRangeList
@@ -211,28 +211,6 @@ object GpuOrcScan extends Arm {
         } else {
           downCastAnyInteger(col, toDt)
         }
-
-      // bool to float, double(float64)
-      case (DType.BOOL8, DType.FLOAT32 | DType.FLOAT64) =>
-        col.castTo(toDt)
-
-      // bool to string
-      case (DType.BOOL8, DType.STRING) =>
-        withResource(col.castTo(toDt)) { casted =>
-          // cuDF produces "ture"/"false" while CPU outputs "TRUE"/"FALSE".
-          casted.upper()
-        }
-
-      // integer to float, double(float64), string
-      case (DType.INT8 | DType.INT16 | DType.INT32 | DType.INT64,
-      DType.FLOAT32 | DType.FLOAT64 | DType.STRING) =>
-        col.castTo(toDt)
-
-      // {bool, integer types} to timestamp(micro seconds)
-      case (DType.BOOL8 | DType.INT8 | DType.INT16 | DType.INT32 | DType.INT64,
-      DType.TIMESTAMP_MICROSECONDS) =>
-        OrcCastingShims.castIntegerToTimestamp(col, fromDt)
-
       // TODO more types, tracked in https://github.com/NVIDIA/spark-rapids/issues/5895
       case (f, t) =>
         throw new QueryExecutionException(s"Unsupported type casting: $f -> $t")
@@ -253,32 +231,18 @@ object GpuOrcScan extends Arm {
       // Align with what CPU does.
       return false
     }
-    val toType = to.getCategory
     from.getCategory match {
       case BOOLEAN | BYTE | SHORT | INT | LONG =>
-        toType match {
-          case BOOLEAN | BYTE | SHORT | INT | LONG | FLOAT | DOUBLE | STRING |
-               TIMESTAMP => true
-          // BINARY and DATE are not supported by design.
-          // The 'to' type (aka read schema) is from Spark, and VARCHAR and CHAR will
-          // be replaced by STRING. Meanwhile, cuDF doesn't support them as output
-          // types, and also replaces them with STRING.
-          // TIMESTAMP_INSTANT is not supported by cuDF.
+        to.getCategory match {
+          case BOOLEAN | BYTE | SHORT | INT | LONG => true
           case _ => false
         }
       case VARCHAR =>
-        toType == STRING
-      case _ => false
+        to.getCategory == STRING
+      // TODO more types, tracked in https://github.com/NVIDIA/spark-rapids/issues/5895
+      case _ =>
+        false
     }
-  }
-
-  /**
-   * Test whether if a * b will cause Long-overflow.
-   * In Math.multiplyExact, if there is an integer-overflow, then it will throw an
-   * ArithmeticException.
-   */
-  def testLongMultiplicationOverflow(a: Long, b: Long) = {
-    Math.multiplyExact(a, b)
   }
 }
